refactor: Move Queue Handling
This moves the queue handling out of common into a more relevantly named section of the codebase.
